Remove focus flag and `ImguiRenderLoopFlags` (#123)
There is no need for that "focused" flag as it was only ever updated by a WM_ACTIVATE event. No other events would arrive when the windows is inactive anyway, so there is no way of getting spurious inputs, which the flag was originally intended for anyway, back when the input loop would use GetAsyncKeyState.